loadgen/requirements.txt
aio-pika
pyyaml
nltk
pymongo